+++ Anchor.toml
@@ -9,4 +9,4 @@ cluster = "localnet"
 wallet = "/Users/alan/.config/solana/id.json"
 
 [scripts]
-test = "ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts"
+test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.ts"
