+++ programs/anchor-escrow-new/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod state;
+
 use anchor_lang::prelude::*;
 use anchor_spl::{
     associated_token::AssociatedToken,
@@ -137,23 +139,7 @@ pub mod anchor_escrow {
     }
 }
 
-#[account]
-pub struct Offer {
-    // We store the offer maker's key so that they can cancel the offer (we need
-    // to know who should sign).
-    pub maker: Pubkey,
 
-    // What kind of tokens does the offer maker want in return, and how many of
-    // them?
-    pub taker_mint: Pubkey,
-    pub taker_amount: u64,
-
-    // When the maker makes their offer, we store their offered tokens in an
-    // escrow account that lives at a program-derived address, with seeds given
-    // by the `Offer` account's address. Storing the corresponding bump here
-    // means the client doesn't have to keep passing it.
-    pub escrowed_maker_tokens_bump: u8,
-}
 
 #[derive(Accounts)]
 #[instruction(escrowed_maker_tokens_bump: u8)]
